feat: хранение расстояния в метрах, то есть Int-ом

+++ src/main/kotlin/domain/training/Training.kt
@@ -45,7 +45,7 @@ sealed class Training(
     /**
      * Информация о беге
      */
-    class Jogging(date: LocalDate, duration: Duration, val distance: Double) : Training(
+    class Jogging(date: LocalDate, duration: Duration, val distance: Int) : Training(
         date = date,
         duration = duration
     ) {
